course 5.56

+++ Services/CharacterService/CharacterService.cs
@@ -27,9 +27,12 @@ namespace udemy1.Services.CharacterService
         {
             var ServiceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = _mapper.Map<Character>(newCharacter);
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+
             _context.Characters.Add(character);
             await _context.SaveChangesAsync();
             ServiceResponse.Data = await _context.Characters
+            .Where(c => c.User.Id == GetUserId())
             .Select(c => _mapper.Map<GetCharacterDto>(c))
             .ToListAsync();
             return ServiceResponse;
@@ -41,10 +44,21 @@ namespace udemy1.Services.CharacterService
 
             try
             {
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
-                response.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                Character character = await _context.Characters
+                .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if (character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
+                    response.Data = _context.Characters
+                    .Where(c => c.User.Id == GetUserId())
+                    .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Message = "Character not found!";
+                }
             }
             catch (Exception ex)
             {
@@ -68,7 +82,8 @@ namespace udemy1.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             var ServiceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var dbCharacter = await _context.Characters
+            .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
             ServiceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             return ServiceResponse;
 
@@ -81,18 +96,26 @@ namespace udemy1.Services.CharacterService
             try
             {
                 Character character = await _context.Characters
+                .Include(c => c.User)
                 .FirstOrDefaultAsync(c => c.Id == updateCharacter.Id);
-                _mapper.Map(updateCharacter, character);
 
-                // character.Name = updateCharacter.Name;
-                // character.HitPoints = updateCharacter.HitPoints;
-                // character.Strength = updateCharacter.Strength;
-                // character.Defense = updateCharacter.Defense;
-                // character.Intelligence = updateCharacter.Intelligence;
-                // character.Class = updateCharacter.Class;
+                if (character.User.Id == GetUserId())
+                {
+                    character.Name = updateCharacter.Name;
+                    character.HitPoints = updateCharacter.HitPoints;
+                    character.Strength = updateCharacter.Strength;
+                    character.Defense = updateCharacter.Defense;
+                    character.Intelligence = updateCharacter.Intelligence;
+                    character.Class = updateCharacter.Class;
 
-                await _context.SaveChangesAsync();
-                response.Data = _mapper.Map<GetCharacterDto>(character);
+                    await _context.SaveChangesAsync();
+                    response.Data = _mapper.Map<GetCharacterDto>(character);
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Message = "Character not found.";
+                }
             }
             catch (Exception ex)
             {
